move and rename test flows
